+++ CHANGELOG.md
@@ -7,11 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0] - 2026-07-09
+
+### Added
+
+- Added underwater night vision support so underwater visibility is no longer restricted. @zimuya4153
+
+### Changed
+
+- Adapted the mod for LeviLamina 26.20.05. @zimuya4153
+
 ## [1.0.0] - 2026-07-08
 
 ### Added
 
 - Initial release @zimuya4153
 
-[Unreleased]: https://github.com/MiracleForest/iInfiniteNightVision/compare/v1.0.0...HEAD
+[Unreleased]: https://github.com/MiracleForest/iInfiniteNightVision/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/MiracleForest/iInfiniteNightVision/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/MiracleForest/iInfiniteNightVision/releases/tag/v1.0.0